requirements.txt
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
bottle==0.12.18
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colored==1.4.2
conan==1.30.2
deprecation==2.0.7
distro==1.5.0
fasteners==0.15
future==0.18.2
idna==2.10
isort==5.6.4
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
mypy-extensions==0.4.3
node-semver==0.6.1
packaging==20.4
patch-ng==1.17.4
pathspec==0.8.0
pluginbase==1.0.0
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
six==1.15.0
terminaltables==3.1.0
toml==0.10.1
tqdm==4.51.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
wrapt==1.12.1
